+++ libdispatch/dfilter.c
@@ -18,7 +18,6 @@
 #include "netcdf.h"
 #include "netcdf_filter.h"
 #include "ncdispatch.h"
-#include "nc4internal.h"
 #include "nclog.h"
 #include "ncglobal.h"