+++ libdispatch/dinstance_intern.c
@@ -16,10 +16,11 @@ Currently two operations are defined:
 #include <string.h>
 #include <assert.h>
 #include "netcdf.h"
-#include "nc4internal.h"
 #include "nc4dispatch.h"
 #include "ncoffsets.h"
 #include "ncbytes.h"
+#include "ncglobal.h"
+#include "nc4internal.h"
 
 #undef REPORT
 #undef DEBUG